refactor(settings): extract EditorSection to sections/
Moves the Editor tab — default note type, line-height picker, and
writing-assistance toggles (spell check, auto-capitalize, word count,
tags) — into its own file under src/components/settings/sections/.

No behavior changes.
